requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
sqlite3
torch==2.0.1
transformers==4.35.2
requests==2.31.0
pandas==2.1.3
plotly==5.18.0
bcrypt==4.1.1
